backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.13.1
openai
requests
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
twilio==9.8.0
